chore(lints): move tests into their own file, fix the lints the standard clippy now brings up
