+++ .gitignore
@@ -11,3 +11,5 @@ Cargo.lock
 
 # IDE files
 .idea
+
+.vscode
