+++ src/instance.rs
@@ -1,3 +1,10 @@
+#![expect(
+    clippy::derived_hash_with_manual_eq,
+    clippy::suspicious_op_assign_impl,
+    clippy::suspicious_arithmetic_impl,
+    reason = "TODO: fix this or add an explanation"
+)]
+
 use std::cmp::Ordering;
 use std::fmt;
 use std::hash::{Hash, Hasher};
@@ -27,7 +34,7 @@ macro_rules! fmt_debug {
 }
 
 /// A boolean value.
-#[derive(Clone, Copy, Hash, Ord, PartialEq, PartialOrd)]
+#[derive(Clone, Copy, Hash, Ord, PartialEq, PartialOrd, Default)]
 pub struct Boolean(bool);
 
 impl GetSize for Boolean {
@@ -112,12 +119,6 @@ impl NumberInstance for Boolean {
 impl RealInstance for Boolean {
     const ONE: Self = Boolean(true);
     const ZERO: Self = Boolean(false);
-}
-
-impl Default for Boolean {
-    fn default() -> Boolean {
-        Self(false)
-    }
 }
 
 impl From<bool> for Boolean {
@@ -534,7 +535,7 @@ impl Sum for Complex {
     fn sum<I: Iterator<Item = Self>>(iter: I) -> Self {
         let mut sum = ComplexType::Complex.zero();
         for i in iter {
-            sum = sum + i;
+            sum += i;
         }
         sum
     }
@@ -599,7 +600,7 @@ impl Product for Complex {
                 return zero;
             }
 
-            product = product * i;
+            product *= i;
         }
         product
     }
@@ -941,7 +942,7 @@ impl Sum for Float {
     fn sum<I: Iterator<Item = Self>>(iter: I) -> Self {
         let mut sum = FloatType::Float.zero();
         for i in iter {
-            sum = sum + i;
+            sum += i;
         }
         sum
     }
@@ -1017,7 +1018,7 @@ impl Product for Float {
                 return zero;
             }
 
-            product = product * i;
+            product *= i;
         }
         product
     }
@@ -1153,11 +1154,7 @@ impl CastFrom<Complex> for Float {
 impl CastFrom<Float> for Boolean {
     fn cast_from(f: Float) -> Boolean {
         use Float::*;
-        let b = match f {
-            F32(f) if f == 0f32 => false,
-            F64(f) if f == 0f64 => false,
-            _ => true,
-        };
+        let b = !matches!(f, F32(0f32) | F64(0f64));
 
         Boolean(b)
     }
@@ -1237,8 +1234,8 @@ impl Collate for FloatCollator {
             order
         } else {
             match (left, right) {
-                (Float::F32(l), Float::F32(r)) => self.f32.cmp(l.into(), r.into()),
-                (Float::F64(l), Float::F64(r)) => self.f64.cmp(l.into(), r.into()),
+                (Float::F32(l), Float::F32(r)) => self.f32.cmp(l, r),
+                (Float::F64(l), Float::F64(r)) => self.f64.cmp(l, r),
                 (l, r) => self.f64.cmp(&(*l).cast_into(), &(*r).cast_into()),
             }
         }
@@ -1372,13 +1369,7 @@ impl CastFrom<Float> for Int {
 impl CastFrom<Int> for Boolean {
     fn cast_from(i: Int) -> Boolean {
         use Int::*;
-        let b = match i {
-            I16(i) if i == 0i16 => false,
-            I32(i) if i == 0i32 => false,
-            I64(i) if i == 0i64 => false,
-            _ => true,
-        };
-
+        let b = !matches!(i, I16(0i16) | I32(0i32) | I64(0i64));
         Boolean(b)
     }
 }
@@ -1513,7 +1504,7 @@ impl Sum for Int {
     fn sum<I: Iterator<Item = Self>>(iter: I) -> Self {
         let mut sum = IntType::Int.zero();
         for i in iter {
-            sum = sum + i;
+            sum += i;
         }
         sum
     }
@@ -1587,7 +1578,7 @@ impl Product for Int {
                 return zero;
             }
 
-            product = product * i;
+            product *= i;
         }
         product
     }
@@ -1871,13 +1862,7 @@ impl CastFrom<Int> for UInt {
 impl CastFrom<UInt> for bool {
     fn cast_from(u: UInt) -> bool {
         use UInt::*;
-        match u {
-            U8(u) if u == 0u8 => false,
-            U16(u) if u == 0u16 => false,
-            U32(u) if u == 0u32 => false,
-            U64(u) if u == 0u64 => false,
-            _ => true,
-        }
+        !matches!(u, U8(0u8) | U16(0u16) | U32(0u32) | U64(0u64))
     }
 }
 
@@ -2015,7 +2000,7 @@ impl Sum for UInt {
     fn sum<I: Iterator<Item = Self>>(iter: I) -> Self {
         let mut sum = UIntType::UInt.zero();
         for i in iter {
-            sum = sum + i;
+            sum += i;
         }
         sum
     }
@@ -2093,7 +2078,7 @@ impl Product for UInt {
                 return zero;
             }
 
-            product = product * i;
+            product *= i;
         }
         product
     }
